+++ python/eyes_robotframework/CHANGELOG.md
@@ -1,6 +1,8 @@
 [vNext]
 ### Added
 - Coded region keywords now support optional padding and region_id arguments [Trello 2977](https://trello.com/c/lYJnhwCz)
+### Updated
+- Distribute code in form of universal wheel package [Trello 2962](https://trello.com/c/0HdBnfAK)
 
 ## [5.7.0] - 2022-07-07
 ### Added